Projectile netcode update

+++ Assets/_Project/Scripts/ScriptableObjectScripts/ArcanaSO.cs
@@ -6,6 +6,6 @@ public class ArcanaSO : AttackSO
     [Header("Arcana")]
     public bool airOk;
     public bool reversal;
-    public int projectileSpeed;
+    public float projectileSpeed;
     public bool projectileDestroyOnHit;
 }
